+++ app/models/carto/user.rb
@@ -79,6 +79,7 @@ class Carto::User < ActiveRecord::Base
   has_many :received_notifications, inverse_of: :user
 
   has_many :api_keys, inverse_of: :user
+  has_many :user_multifactor_auths, class_name: Carto::UserMultifactorAuth
 
   has_many :oauth_apps, inverse_of: :user, dependent: :destroy
   has_many :oauth_app_users, inverse_of: :user, dependent: :destroy