+++ config/routes.rb
@@ -636,6 +636,12 @@ CartoDB::Application.routes.draw do
                                   controller: :received_notifications,
                                   constraints: { id: UUID_REGEXP }
       end
+
+      # Multi-factor authentication
+      resources :multifactor_auths, only: [:create, :destroy], constraints: { id: /[^\/]+/ }
+      scope 'multifactor_auths/:id' do
+        post 'validate_code' => 'multifactor_auths#validate_code', as: :validate_multifactor_auth
+      end
     end
 
     scope 'v2/' do